02-workflow-orchestration/practice_airflow/requirements.txt
pandas 
sqlalchemy
psycopg2-binary
apache-airflow[google]
pyarrow